remove --ip alias for --bind
The addition --ip option was _intended_ to be added for compatibility
with the v3 replicator option of that name, but was based on a too-quick
reading of the v3 code.  The v3 --ip option is *not* merely a different
name for the v4 --bind option, but rather intended as a means of
filtering incoming requests.

The intent is to add the v3 --ip and --alias options before "beta",
but this commit is just removing the inappropriate meaning of --ip
to avoid any further confusion.